areas, ftn2rfc: add support for plain headers
RFC5335 allow UTF-8 in headers, so make it possible to leave the
headers untouched by qp/b64 encoder.

It's up to the sysop to define charsets correctly (to utf-8) in
fidogate.conf (DefaultCharset, NetMailCharset) or in areas.

It also leaves long lones. In case if it's critical the encoder can
be extented by plain splitter (implement encode() method, all the
infrastructure is in place).

Config: add NetMailHeadersPlain option. Add -nh areas option.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>